AnimationUtils: Move new helpers into GLTFExporter.
